+++ components/confluence-sync/src/errors/InvalidSyncModeException.ts
@@ -1,0 +1,13 @@
+// SPDX-FileCopyrightText: 2025 Telefónica Innovación Digital and contributors
+// SPDX-License-Identifier: Apache-2.0
+
+import { SyncModes } from "../ConfluenceSyncPages.types";
+
+export class InvalidSyncModeException extends Error {
+  constructor(syncMode: string, options?: ErrorOptions) {
+    super(
+      `Invalid sync mode "${syncMode}". Please use one of "${SyncModes.FLAT}", "${SyncModes.TREE}" or "${SyncModes.ID}"`,
+      options,
+    );
+  }
+}