+++ components/confluence-sync/src/errors/RootPageRequiredException.ts
@@ -1,26 +1,15 @@
 // SPDX-FileCopyrightText: 2024 Telefónica Innovación Digital and contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { ConfluenceInputPage } from "../ConfluenceSyncPages.types";
 import { SyncModes } from "../ConfluenceSyncPages.types";
-import { getPagesTitlesCommaSeparated } from "../support/Pages";
 
 export class RootPageRequiredException extends Error {
-  constructor(
-    mode: SyncModes,
-    pagesWithoutId?: ConfluenceInputPage[],
-    options?: ErrorOptions,
-  ) {
+  constructor(mode: SyncModes, options?: ErrorOptions) {
     /* istanbul ignore else */
     if (mode === SyncModes.TREE) {
       super("rootPageId is required for TREE sync mode", options);
     } else if (mode === SyncModes.FLAT) {
-      super(
-        `rootPageId is required for FLAT sync mode when there are pages without id: ${getPagesTitlesCommaSeparated(
-          pagesWithoutId as ConfluenceInputPage[],
-        )}`,
-        options,
-      );
+      super("rootPageId is required for FLAT sync mode", options);
     } else {
       /* istanbul ignore next */
       super("Unknown sync mode", options);